+++ lib/akismet.dart
@@ -5,6 +5,9 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 import 'package:http/http.dart' as http;
+import 'io.dart'
+  if (dart.library.html) 'src/io/browser.dart'
+  if (dart.library.io) 'src/io/vm.dart';
 
 part 'src/author.dart';
 part 'src/blog.dart';
@@ -12,4 +15,4 @@ part 'src/client.dart';
 part 'src/comment.dart';
 
 /// The version number of this package.
-const String _version = '3.1.0';
+const String _version = '3.2.0';