+++ lib/io.dart
@@ -1,0 +1,7 @@
+/// Provides the I/O support.
+library akismet.io;
+
+import 'package:http/http.dart' as http;
+
+/// A reference to the HTTP client.
+final http.Client httpClient = null;