+++ lib/src/client.dart
@@ -76,7 +76,7 @@ class Client {
       ..headers[HttpHeaders.USER_AGENT] = userAgent;
 
     _onRequest.add(request);
-    var response = await http.post(request.url, body: request.bodyFields, headers: request.headers);
+    var response = await httpClient.post(request.url, body: request.bodyFields, headers: request.headers);
     _onResponse.add(response);
 
     if ((response.statusCode / 100).truncate() != 2) throw new http.ClientException('An error occurred while querying the end point.', endPoint);