+++ lib/src/io/browser.dart
@@ -1,0 +1,5 @@
+import 'package:http/http.dart' as http;
+import 'package:http/browser_client.dart';
+
+/// A reference to the HTTP client.
+final http.Client httpClient = new BrowserClient();