+++ lib/src/io/vm.dart
@@ -1,0 +1,4 @@
+import 'package:http/http.dart' as http;
+
+/// A reference to the HTTP client.
+final http.Client httpClient = new http.IOClient();